wip: some reorganization, not sure how robust it is yet
