+++ src/tellus/core/__init__.py
@@ -1,0 +1,3 @@
+"""Core functionality for the Tellus CLI."""
+
+__all__ = []
