+++ src/tellus/location/__init__.py
@@ -1,0 +1,15 @@
+"""Location management for Tellus."""
+
+from .location import (
+    Location,
+    LocationKind,
+    LocationExistsError,
+    create_location_handlers,
+)
+
+__all__ = [
+    "Location",
+    "LocationKind",
+    "LocationExistsError",
+    "create_location_handlers",
+]
